+++ src/ScottPlot5/ScottPlot5 Cookbook/Recipes/PlotTypes/Signal.cs
@@ -27,7 +27,7 @@ internal class Signal : RecipePageBase
             sig2.Data.YOffset = .25;
             sig2.Label = "Offset";
 
-            myPlot.Legend();
+            myPlot.Legend.IsVisible = true;
         }
     }
 }