+++ src/ScottPlot5/ScottPlot5/Legends/StandardLegend.cs
@@ -1,8 +1,6 @@
-﻿using ScottPlot;
+﻿namespace ScottPlot.Legends;
 
-namespace ScottPlot.Legends;
-
-public class StandardLegend : ILegend
+public class StandardLegend
 {
     public bool IsVisible { get; set; } = false;
     public Alignment Alignment { get; set; } = Alignment.LowerRight;