+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Plot: Benchmark is now a user-customizable plottable and `Plot.ShowBenchmark` is now `Plot.Benchmark.IsVisible` (#2961)
 * Grid: Improve support for custom line styles (#2904) _Thanks @minjjKang_
 * Pie: Improve appearance of slice labels in the legend (#2894, #2852) _Thanks @zy1075984_
+* Legend: Replaced `List<ILegend>` with a simple `Legend` object with an `IsVisible` property (#2792)
 
 ## ScottPlot 4.1.68 (in development)
 * Axis: Added `IsReverse` property to let users invert the orientation of an axis (#2958) _Thanks @HandsomeGoldenKnight_